Big Endian detection: Un-break on BSD et al (alternate approach)
This is an alternate approach to what SODA is doing in pull request
number #34. This steals an idea Liam R. Howlett
<Liam.Howlett@Oracle.com> (git mailing list
<20170627162808.p7sj3jpwum7n3hq2@oracle.com>) of using macro string
concat as a hack to see if the macro is defined to an empty value.

SODA's patch is simply:

    #if defined(_BYTE_ORDER) && \
    +    defined(_BIG_ENDIAN) && defined(_LITTLE_ENDIAN) && \
    +    _BYTE_ORDER == _BIG_ENDIAN

In some ways this is a lot simpler than this, but the advantage of
this is that we can fold the GCC condition as well into this "let's
define bigendian/byte order values we see".

I ran out of time, so this patch doesn't actually do that, but this
patch allows us to move back to the sort of check we had before my
56ab30c ("Rewrite Big Endian detection (again)", 2017-06-27) without
breaking on systems like Solaris that have empty values.